ref: move code highlight to python side
In order to implement a nice side-by-side highligted diff view we need
a diff library, the two file raw source code and extra abilities to
controle the generated html.

Using highlight.js proved to be too much cubersome as the python side
had to send the raw file and let javascript does everything, run a diff
algorithm, insert the new lines when needed, ensure the line numbering
is sync between the two files and highlight the source code.

As no JS dev took the challenge, this commit move all the highlighting
to python side. The library Pygments is a python library used by many
projects to hightlight source code. It is hackable and can produce HTML.

Bumping MAJOR as this is not retrocompatible.

Related to #7
Related to #8